requirements.txt
discord.py
python-dotenv
newsapi-python
groq
